Fix missing Wikibase CSS on undo and diff pages
This fixes specifically the problem of language fallback indicators
being shown when they’re supposed to be hidden, but more generally we
should probably include the Wikibase CSS if we’re showing rendered
entity links. (Maybe all of wikibase.common is a bit much for this case,
but this is not the right moment to split it up, IMHO.)

The helper function in EditEntityActionTest is expanded a bit to be able
to test that wikibase.common is included in the ResourceLoader modules.

Bug: T267502
Change-Id: Ibd4cdd718eb0af3d9453786a66b2efc48862fc5b
(cherry picked from commit 3479875548cbe285541dd75d48274e9c10116aaf)