EditEntityActionTest: don’t call overrideMwServices()
overrideMwServices() is a fairly heavy-handed method – it completely
replaces the service container, which can cause issues if anything else
is holding a reference to the service container (as proposed in
I57df20e94d). Its documentation says that it “should not be used to
reset services in between stages of a test”, and advises to call
resetServices() instead, so let’s do that. (The first call seems to have
been unnecessary, as far as I can tell, and is removed without
replacement. Maybe it was only needed during a transition phase of the
PermissionManager changes – see change Id274f240d6, commit 88be000489.)

Change-Id: I0b09708391434367ac37b5a3e19ba4531d3307f9