repo: move data+entity type definitions to service wiring
This first iteration is *only* meant to move the initialization code out
of WikibaseRepo::newInstance(). The “lifecycle” of the two definitions
services doesn’t change yet: they’re still stored as members of the
WikibaseRepo singleton, and should not (yet) be accessed in the service
container directly. Ultimately, they and all other Wikibase services
will probably live only in the service container (and the WikibaseRepo
class will only be an accessor for the service container, without a
singleton instance of its own), but we’re not there yet, and for the
time being, keeping the definitions lifetime tied to the singleton seems
like the safest way forward.

Bug: T265984
Change-Id: I9abcaa35b32b52f1fed5b04f84b9e6f3596e8de4