Use MediaWikiIntegrationTestCase, not MediaWikiTestCase
This is not to say that every existing use is in fact an integration
test, it just consistently uses the primary name
(MediaWikiIntegrationTestCase) so that developers are easily aware of
the feature set available inside of their tests (I was not aware that
MediaWikiTestCase is an alias and thought there is a difference). See
MediaWiki's 214750d for the backstory.

Also consistently adds `use` to the file heads as was the case with the
vast majority of the tests using already MediaWikiIntegrationTestCase so
far.

Change-Id: I1f6823bd6cf554a663ccf016d56bd82d2b3c8086