Only load termbox on Item and Property pages
Don’t attempt to load the new termbox on other entity types, it doesn’t
work there.

This can only be tested if at least one other entity type is available
on the local wiki, so the test data provider is refactored to use
'yield', which makes it easier to conditionally add a test case.

Bug: T264893
Change-Id: I204c8930dcb7a1fdf68b16e525d1b8cc455702e3