Calculate page props on-the-fly during RDF dump
The RdfBuilder now takes an EntityContentFactory instead of an
EntityTitleLookup, and addEntityPageProps() is called with a full
EntityDocument instead of an EntityId, so that it can pass that entity
into the EntityContentFactory to get an EntityContent and ask that for
the page props, instead of looking them up in a PageProps store. This
means that we make the content object recalculate the page props instead
of looking them up, guaruanteeing that we see fresh data (the page_props
table is updated asynchronously, so previously the data could be
slightly stale) that also matches the rest of the entity data (in RDF
for old revisions, the page props were previously completely unrelated
to the rest of the data, and would always reflect the props of the
latest revision).

Several other classes in the pipeline also need to be updated to pass
around an EntityContentFactory instead of an EntityTitleLookup;
fortunately, EntityContentFactory implements EntityTitleLookup, so in
contexts where the latter interface is still needed, we can use the same
object for both roles.

In RdfBuilderTest there is a slight paradigm change: it used to override
the RdfBuilder’s PageProps instance with a setter on the class, but I
think it’s cleaner to leave test-only setters out of the class and use
TestingAccessWrapper instead to directly set the private field. (The
test is dedicated to that specific class, so I think binding to its
implementation in this way is acceptable.)

Bug: T145712
Change-Id: I1a0e3962eaa5b8c644b9c70f6bf656a7cd0fa2c6
Depends-On: I3233756bfe7482e99e3e575da2c8727b1938d99a
Depends-On: Ic90bf621bda9787fffe890c34a3bc6b8cd3452a0