Delete sitelink when moving to excluded namespace without redirect
When a page was being moved to an excluded namespace, the hook handler
would previously abort immediately, leaving the item sitelink to the
moved page as it is. That’s fine if a redirect was left behind (it’s
better to link to the redirect than to nothing at all, if we don’t want
to link to the excluded namespace), but if the creation of the redirect
was suppressed, then we want the repo sitelink to be deleted.
Fortunately, this is fairly straightforward to implement thanks to page
moves and deletions being handled very similarly: we can reuse the same
code as for the case where the page was actually deleted. (We also
extract an applyUpdateRepo() method to reduce duplicate code.)

While we’re at it, also modernize the tests a bit, with named arguments
in the data provider and better type declarations.

Also, set the dynamic wikibasePushedMoveToRepo property on the parameter
we received ($newLinkTarget), not the value we created ourselves ($nt).
This fix really belongs with T268135, but it becomes necessary here
because in the “excluded namespace” branch, $nt does not exist, and it
doesn’t really hurt to fix this now either.

Bug: T261275
Change-Id: I5f6cd230d3f407a795ca08df3ce1959f66118825
Co-Authored-By: Noa Rave <noa.rave@wikimedia.de>