Rename page props-related identifiers for clarity
The page properties and page property *definitions* are very different
and should not have confusingly similar names. (Prior to change
I1a0e3962ea, this was even worse, since we also had the PageProps class,
which is the service to *access* page properties. The class still
exists, of course, but at least we no longer use it here.)

Bug: T145712
Change-Id: Id0c26d8178a86cdcc8d23854de5c12adb877a3ca