Log+skip bad page IDs when reading/writing entity usage
The eu_page_id should never be 0 (or negative, though I assume that the
bug we’re seeing is that it’s 0), but apparently that occasionally
happens, leading to an exception from the PageEntityUsages constructor.
Avoid this by checking when we’re encountering a page ID whether it’s
valid or not (and logging the occurrence), but then skipping the bad
rows, so that the rest of the request can keep running.

The logger is not injected because properly injecting it looks like it
would require changes to at least three more classes, and I don’t think
it’s worth it, since this logging will hopefully be temporary (the
logged warnings should tell us where the bad page IDs come from, we’ll
fix them, and then turn the warnings into exceptions).

Logging 'exception' => new RuntimeException() is a common pattern in
MediaWiki to include a stack trace in the structured log data. (There
are some variations, but RuntimeException with no message seems to be
the most common version according to code search.)

While touching the class, also add lots of type declarations and declare
strict types. (This makes one test obsolete.)

Bug: T264929
Change-Id: Ib984c1948ecd93d930ae4ade29c2063e4ea2a334