SECURITY: Add job to purge entity data on page deletion
Make EntityDataPurger handle a second hook, for when an entire page is
deleted; however, we don’t handle the deletion directly in the hook
handler. Instead, define a new job class to purge the entity data URLs
for all archived revisions of a deleted page, and make the hook handler
push that job. Deferring the purging to a job ensures that the main
transaction (archiving the rows) has committed and we can therefore read
the archived rows from a replica (though we still need to wait for
replication, in case the job runs immediately), and it’s probably also a
good idea in general to not read an unbounded number of archive rows in
the web request that is deleting the page.

The dependency injection pattern of PurgeEntityDataJob is mainly based
on ChangeVisibilityNotificationJob; the practice of using the standard
'namespace' and 'title' params even for a deleted page follows
DeletePageJob (MediaWiki core).

An additional feature not implemented here would be to pass the hook
handler’s $archivedRevisionCount into the job parameters, and have the
job log a warning if it processed a different number of archive rows.
However, I’ll leave that for later, once this patch becomes public.

(Note: the tests which use LBFactorySingle require MediaWiki change
I6cb62cfeef to work properly. I decided against adding a Depends-On
footer, since it’s a test-only dependency and I don’t want to risk
confusing deployers. If the MediaWiki change is not merged for some
reason, a workaround is to use Wikibase’s FakeLBFactory instead.)

Bug: T260349
Change-Id: I8f88178721c6c941530a03ca431725820a24d468