Log more messages in LoggingIdGenerator
Add a debug-level message just before calling the inner ID generator,
and another one if that throws an exception. This should make the class
more useful in combination with other IdGenerator implementations.

Specifically, if we add a rate limit to entity ID generation, I would
implement that as another decorating IdGenerator implementation similar
to LoggingIdGenerator, and in that case I think it would make sense to
make LoggingIdGenerator wrap RateLimitingIdGenerator, not the other way
around – that means that, with this change, we would still get some log
message if the rate limit is hit (but only on DEBUG level, not INFO,
since no ID was actually “wasted”).

Bug: T268625
Change-Id: If410eeed56daf32839f16157625125d207cb83b1