Move WikibaseRepo constructor params to services
This moves both the Settings and the EntitySourceDefinitions out of the
WikibaseRepo instance into the service container. They have to be moved
at the same time, since the settings name the default entity source, so
if only one of the two is moved to services, then some tests will start
failing because the default entity source in the custom settings is not
in the default entity sources, or vice versa.

A handful of tests are updated to set the services for the Settings and
the EntitySourceDefinitions instead of passing them to the WikibaseRepo
constructor; almost all other users of WikibaseRepo already relied upon
the global state (i.e. the class was very rarely directly instantiated),
so not much changes for them. (I had initially considered having a copy
of the services in the WikibaseRepo instance, and checking whether they
matched the service container’s version, but it turns out we don’t need
that, as far as I can tell.)

Calls to getSettings() and getEntitySourceDefinitions() should be fixed
to call the methods statically, but I’m doing that separately, in order
to keep this commit smaller and easier to review.

Bug: T270479
Change-Id: I594d5a59f0866fa3823922f9aa09700cc677b29f