LocalizedTextProvider: add getEscaped method
The previous interface description of LocalizedTextProvider only
contained a method to get unescaped text. This seems impractical given
that in most cases it would need to be escaped by the consuming code
since it can't fully trust the source of the localized texts. It also
makes it awkward to get safe, unescaped parameters containing markup
into the text without also leaving the message itself unescaped.

This patch adds an additional `getEscaped()` method to the interface for
the common purpose to get escaped localized text, and to provide a safe
way to use unescaped "raw" parameters, which will be added in a later
patch.

Bug: T267023
Change-Id: I381403bcb969bc7db40da8b3d1e6c6e405b012f5