Replace wikibase-codesniffer with mediawiki-codesniffer
This copies the rules and exclusions from the wikibase-codesniffer
ruleset.xml file and adds a handful more that were moved to
mediawiki-codesniffer but need to be enabled explicitly. Autofixable
code style issues are fixed as well, as well as a handful of manual “no
fully-qualified names” fixes; some more not autofixable rules are
temporarily excluded and will be fixed in separate changes.

Bug: T268827
Change-Id: Icadb4b213ffa77e5580a180d0268ed98ea73769a