Override loaded extensions better in unit test
I don’t know how this test passes in CI, but locally, I get an error
when a FederatedProperties hook handler tries to access the repo
settings, but WikibaseSettings complains that WikibaseRepository is not
a loaded extension. Instead of blindly overriding the extension
registry’s list of loaded extension with 'Wikibase' (two problems:
'Wikibase' is not a registered extension name to begin with, and the
array is also supposed to have extension name *keys* and credits arrays
as values, not numerically indexed extension name values), intersect the
list with only the core Wikibase extensions, client and repo.

Change-Id: Idd0233d2476f69e55888bdba12eb0fe2c108f88e