Change Title::newFromText to newFromTextThrow in tests
Title::newFromTextThrow() throws a MalformedTitleException if the title
cannot be parsed; newFromText() catches that exception and returns null
instead. However, almost none of our tests expect newFromText() to ever
fail; if it unexpectedly returns null, one of three things will happen:

1. We get a PHP error when we try to call a method on the null title
   (e.g. resetArticleId()).
2. We get an TypeError when we pass the null title into a method that is
   typehinted to accept a non-optional Title.
3. The title silently falls back to $wgTitle.

The third case is the rarest, but also the most worrysome of the three;
I suspect it happens in HtmlPageLinkRendererEndHookHandlerTest(Base) and
makes some of the test failures we see in T269608 more confusing.

In all cases, it’s preferable to throw a MalformedTitleException early
in the test. Only two tests actually expect newFromText() to potentially
return null and handle this case: AffectedPagesFinderTest and
ChangeHandlerTest.

I tested manually that this doesn’t cause any new test failures in repo,
client, or lib. (I have one existing test failure in client, but that
seems unrelated.)

Change-Id: I5344ff004b29541ea983ebb199e95c18f5e9b90f