Add WikibasePingback
We want to introduce a similar pingback functionality similar to
Mediawiki. There has been a discussions about possible ways forward
wheather to extend the current mw pingback (T269674) or copy paste the
existing one and adapting it to our needs.

Seeing that we are working in a limited timeframe during the 2021
preparation week and this is expected to be rolled out at the start of
next year the decision to copy the MW mechanism was made.

Change-Id: I56d862fc98e8f54d82b4d55faf9101cda090842e