Update getSettings() calls
A lot of these can be injected now, especially in APIs and special
pages. When introducing new variables, I generally went for the name
$repoSettings rather than just $settings, to make it a bit clearer what
the settings belong to – this is especially important, I think, when
they are injected as in the various factory methods, where readers of
the code can’t easily see where the variable comes from.

Bug: T270479
Change-Id: If426bf5c7151e0a4a1a34c5d9441116579e6d077