Skip 0 page ID in hook handler
If the page ID is 0 for some reason, we can’t really use it, and
eventually the PageEntityUsages constructor will throw an exception. I
think it’s better to skip the whole hook handler in that case. Some
logging is included to possibly gain a better understanding of what
causes this (I added the $linksUpdate->getCauseAction() because it
sounds possibly useful, though admittedly I have no idea what it
actually is).

Bug: T264929
Change-Id: I9fe11fdd83acb3d9ac47e0559926a5802f89874b