EntityHandlerTestCase: don’t hard-code item+property
Instead of hard-coding only the item and property entity types into the
test entity source, iterate over all the entity types (which we already
know) and assign a fake namespace ID to each (and assume they’re all
stored in the main slot). This fixes some lexeme tests that otherwise
fail once we move the entity source definitions to services.

Bug: T270479
Change-Id: Ic269cac262cce2fa37d9eeb3933a88b32b9e0219