Don't assume wbUserSpecifiedLanguages is always there
While there *shouldn't* be a situation where the
wb.getUserLanguages() method is available and called, but the
wbUserSpecifiedLanguages config variable is not, this is
obviously what happens here. There is no guarantee. For example,
it's possible to manually load the ResourceLoader module that
contains this .js file and run it on a page that doesn't have
the config variable. The code should not crash then. There is a
nice fallback already in the code. Just use that.

Bug: T269871
Change-Id: If4dbd7307524eb169a43626deab3675a1dd6ad65