Fix HtmlPageLinkRendererEndHookHandler and reenable test
The test broke because an unrelated unit test started setting the MW_API
constant, which made the hook handler skip itself. Make the hook handler
a bit smarter (if the constant is set to 'TEST', act as if it’s not set)
and then reenable the test.

It would probably be preferable to not make that other test define the
constant, but I’m not sure if that’s easily doable. See the linked task
for details.

Bug: T269608
Change-Id: If6388fe8d47beef444e2b6b693d7dfcc3c75cc8e