Fix unused private members
Most of them either became unused over time (but weren’t removed along
with the last code to use them), or were never used but can clearly be
removed, and are therefore removed. One, MonolingualTextParser’s const
FORMAT_NAME, is meant to be used in exception messages (compare change
Id4bfc429a5, commit bb20eb6b67), and after fixing that it’s not unused
anymore. And one, GetClaimsStatementFilter::FILTER_TYPE, I think ought
to be public, so that each StatementFilter implementation has a public
constant with that name, even if for GetClaimsStatementFilter it’s not
currently used.

(IntelliJ reports unused private class properties and constants as the
same problem, that’s why this commit fixes both.)

Bug: T269881
Change-Id: I321412fe82804f83b9313f3d7f15d7c472d7470b