Change default name of Wikibase-repo-name
The default name of a generic Wikibase installation is currently "Wikidata". This is misleading as the link on the sidebar on a custom Wikibase instance would read "Wikidata Item", but this would not be correct: the item linked would not be on Wikidata, but on the custom Wikibase instance.
So the default label should simply be "Wikibase"

Bug: T267098
Change-Id: Ie4ddf0a280a8c08ad025230dc5283445d9507f46