Change default name of Wikibase-repo-name
The default name of a generic Wikibase installation is currently "Wikidata". This is misleading as the link on the sidebar on a custom Wikibase instance would read "Wikidata Item", but this would not be correct: the item linked would not be on Wikidata, but on the custom Wikibase instance.
So the default label should simply be "Wikibase"

Bug: T267098
Depends-On: Id6dc8c9828b6898164c9a8a237e6e8c42a11c93e
Change-Id: Ie4ddf0a280a8c08ad025230dc5283445d9507f46
(cherry picked from commit 06bec855d072ce48df041458c9d1c3973a626e3c)