WikibaseClientTest: Don't mutate settings of default instance
Tests shouldn't have global side-effects or else they can lead to
flaky behavior depending on the order in which tests are run.

In particular, this particular side-effect was causing parser tests
run after Wikibase to fail when trying to delete articles created for
the parser test: the ArticleDeleteComplete hook in wikibase threw a
'No source configured: test' LogicException.

Bug: T270777
Change-Id: I9eee817abb35f5bd3650d93a11ca208b1e3d8bf0