Handle error when mixing redirects in undo
If the redirect status of the newer endpoint of the range of edits to be
undone is not the same as the redirect status of the latest revision of
the page, then report an error to the user. Otherwise, there may be a
crash later when attempting to apply the diff to the latest revision.

This also edits the comments in ActionTestCase so that the “redirect”
case for test items is no longer described as a “hack” – this “uglyness”
(from change I00dd55c02d, commit c52cc38721) has persisted for over six
years, so it clearly wasn’t going to be replaced with something better
soon, and I don’t think it’s all that bad, really.

Bug: T97146
(cherry picked from commit 6e09365fc0599b283c0998190bc6d2dfbd195c00)
Change-Id: I4e97a2d55e601ff2f46e3ae1db06c90b20cbd55b