Permit restoring redirects to discard entity data
Previously, when an edit that turned an entity from a redirect back into
an ordinary entity was reverted (i.e., when turning an entity into a
redirect again), that undo also had to remove all the data in the
entity, or an error would be thrown. Instead, discard any data left in
the entity after applying the patch (or rather, if any redirect diff can
be applied, don’t even bother applying the non-redirect part).

This also adds some other tests for undo/redirect behavior, which
requires some changes to the testing framework so that we can test edits
that are supposed to leave the entity in a redirect state. Also,
resetTestItem() needs to learn to reset entities that have redirects in
their history every time, even if they weren’t edited, so that the
redirects don’t point at item IDs that no longer exist.

Bug: T97146
Change-Id: Ibac379e7b6538fa6d7fc4d38ff2171688af721c9