Check whether MediaWiki can precisely format a number
MediaWiki no longer formats numbers without loss of precision, so
instead of directly returning Language::formatNum()’s result from the
MediaWikiNumberLocalizer, check if that result parses into the same
number as the input, and otherwise assume that a loss of precision
occurred and fall back to the unformatted number – it’s better to show
an unformatted number than an imprecise one.

The test cases allow both formattings for the inputs that currently
suffer loss of precision, so once MediaWiki learns to format them
accurately (which will hopefully happen eventually), the tests should
continue to pass.

Bug: T268456
Change-Id: I3e5c03f22a834fcaf2b093d9c2e9205f4f40edfe