Update unit conversion script
The most significant change is one in terminology: what the script used
to call “base units”, we now call “coherent units”, since that is the
correct term: units which are not multiples of a “more basic” unit, but
multiples of SI base units with a factor of one. In some sense, those
can be considered “the base” of the unit conversions, but calling them
“base units” is very confusing, since that is the term for exactly seven
SI units – second, metre, kilogram, ampere, kelvin, mole, and candela –
and no other SI units. (This also reduces the degree to which the term
“base” is “overloaded” within the script – the “base URI” was never
related to the “base units”, nor to “Wikibase”.)

The --base-unit-types option is not only renamed, but completely
removed. The correct type to use, Q69197847 (coherent SI unit), is now
hard-coded, because I see little use in making it configurable; of the
previously suggested types, Q223662 (SI base unit) was correct but
incomplete (it’s a subclass of coherent SI unit), whereas Q208469 (SI
derived unit) was flat-out wrong, since units like millimetre are also
SI derived units, but should not be used as the “base” of the unit
conversion: they should not be converted to themselves with a factor of
one. (The situation around these items must have been somewhat different
back in 2016 or 2017, when the script was first written and used, since
it didn’t emit such incorrect conversions then.)

Also, remove the code that skips “exotic” units, i.e. ones that have no
English label. It was already marked with a “TODO: drop this”, and in
fact there are some units on Wikidata that are used in quantity values,
have a conversion to SI, and have no English label (such as Q70422268).

Bug: T270251
Change-Id: Ic182820b5115aafe1080a4efe23bbeb8894de157