Fix unitStorage documentation
The class name was incorrect since change Ibfd2f4c9bd (commit
aea3bd088e), which moved the class to the Units namespace. Also, a
single backslash each time is enough, and for the path to the JSON file,
let’s suggest an absolute path using __DIR__ so it’s not ambiguous what
the path is relative to (this is also what we do in production).

Change-Id: I8857ba37b8f0fbd2c166c9836f714836856a7bae