Add monolog/monolog depencency
We are currently referencing "monolog/monolog" in
SqlChangeDispatchCoordinator for formatting the logs when running
dispatchChanges.php for example.

However this dependency is only mentioned by mediawiki as a dev only
requirement (Not currently a thing in wikibase composer.json)

Installing Wikibase on a fresh mediawiki where you run composer
install with "--no-dev" currently does not give us this dependency and
leads to the change dispatching not working.

Change-Id: I6e50ed210ba167bc6f59f878a259bac0039c5fa5