Make wbeditentity return full entity data
Instead of adding a subset of the entity data to the response, based on
some well-known Wikibase data model interfaces, ask the result builder
to serialize the whole entity and add that to the response.

The impact of this varies by entity type:

* Item: No change, I believe. All aspects of an Item – statements,
  terms, sitelinks – ought to be covered by those interfaces for which
  wbeditentity already emitted data.
* Property: Minor change – the datatype is now included in the response.
* Lexeme: Major, but backwards-compatible change – the lexicographical
  data (including "lemmas", "senses", "forms") is now included in the
  response.
* Sense, Form: similar to Lexeme.
* MediaInfo: Statements are now returned under "statements", as in
  wbgetentities and Special:EntityData for MediaInfo entities. To avoid
  making this a breaking change, we keep a part of the old code, so that
  the response for MediaInfo will temporarily include both "statements"
  and "claims". (It’s worth noting that the two keys don’t contain
  exactly the same data: the JsonDataTypeInjector only adds a "datatype"
  to snaks under "claims", not "statements".) Eventually we intend to
  remove this, and the wbeditentity response for MediaInfo entities will
  then only include "statements" (without "datatype", unless a change is
  made to JsonDataTypeInjector to fix T246809).

This requires a MediaWiki core change (see Depends-On trailer) so that
ApiResult doesn’t complain when ModifyEntity::addToOutput() calls
ResultBuilder::addBasicEntityInformation() to add the "id" and "type" to
the result, which will already have been added by EditEntity.

The entity data serialization is wrapped in a try/catch because it may
throw an exception. This happens, for example, when a request edits an
existing lexeme with the "clear" parameter enabled, but the "data" has
no language or lexical category. In this case, and presumably in other
cases as well, the whole request will fail anyways due to the bad data,
but the result building happens at an earlier stage, so that error may
not have happened yet.

Bug: T160504
Bug: T202725
Bug: T271105
Change-Id: I208c801e1571b0f3a28433e47ae196039c521a67
Depends-On: Icfa6c181666b2293494c67875777d3de8816587a
Depends-On: Ia11ed1df34e09410678eb4e5034a8ae64295a712