Add wikibase-idgenerator rate limit
This rate limit (defaults to the 'create' or 'edit' rate limit –
'create' is not defined by default, but 'edit' is) applies to all
attempts to generate a new numeric entity ID, whether through a special
page, the API, or some other way.

The limit is not (yet) enabled by default – to enable it, set
$wgWBRepoSettings['idGeneratorRateLimiting'] to true. (I assume this is
temporary, and we’ll eventually enable it unconditionally and remove
that setting.)

The rate limit default value is set in the extension callback, since
there’s no natively supported way to inherit a rate limit from another
in extension.json. The pattern of the marker key (&inherit-create-edit)
is meant to resemble the '&can-bypass' marker supported by MediaWiki.

Bug: T272032
Change-Id: Ic701ffd0f7d23516325143be9d9cca072878ce7a