bridge: read externals config from package.json
We were configuring the external libraries in two places: in
package.json, for the @wmde/lib-version-check configuration, and in
vue.config.js, for the webpack configuration. Make vue.config.js read
the packages from the package.json instead, to avoid the two lists
getting out of sync. This is meant to partially follow Pablo’s
suggestion at T271413#6790791 – Pablo also proposed changing the data
structure in the package.json, which I don’t disagree with, but that
would require a new version of @wmde/lib-version-check first and can
still be done later.

Bug: T271413
Change-Id: I160fa3d86982c481cfc1a330ed0745b7e2e07a34