Fix Travis CI build on release branches
If the Wikibase directory already existed, e.g. as an empty directory
(Git submodule) in a MediaWiki release branch tarball from GitHub, then
cp .../Wikibase Wikibase would copy the source Wikibase directory below
that Wikibase directory, resulting in a path .../Wikibase/Wikibase,
which breaks the build. Avoid this with the -T option to make `cp` treat
the target as a normal file, not a directory to copy things into.

Change-Id: If096af4123f8f5034e15c81ced4f1270a4dade0f