Make wb_changes_dispatch.chd_seen unsigned
The column was previously signed, but it references the
wb_changes.change_id column, which is unsigned. Making it unsigned
should prevent it from overflowing before change_id does.

(The postgres patch is empty because PostgreSQL does not support
unsigned integer types, but we still include the file with its comment.)

Bug: T263644
Change-Id: I873c67f3e6a49569be09f2bd9fc60a50d7859053