Let StatementEntityReferenceExtractor be reused
If StatementEntityReferenceExtractor was used multiple times, it would
return the referenced entity IDs of earlier calls later. This is hardly
desirable, so stop doing it and make each call independent instead.

This could be implemented in several ways; the implementation here (keep
$this->entityIds but clear it before returning) is not the nicest, but
at least it’s not a big change from the existing code.

Bug: T273830
Change-Id: I2c0182a72b50646cc55c1d3bfe7ceb55cf057278