Simplify and fix WikiPageEntityMetaDataLookup::selectRevisionInformationById()
Use PageTableEntityQuery instead of directly running a simpler query.
This brings selectRevisionInformationById() in line with
selectRevisionInformationMultiple() & selectLatestRevisionIdsMultiple().

This also subtly changes the behavior of the method. Previously, it only
really looked up the information of the given revision ID and attached
the slot role name of the entity ID to it – there was no check that the
revision belonged to the entity at all. PageTableEntityQuery includes
such a check – not just that the revision matches the entity ID, but
also (for entities stored in non-main slots) that the revision contains
a slot for that entity type. For example, the method will now return
false when called with a MediaInfo ID containing the page ID of a file
page, and a revision ID belonging to that file page, but the file page
has on entity data (no mediainfo slot). (Note that the method had always
been *documented* to return “false if no such entity exists”, so this
change now makes the implementation match the documentation.)

This check in PageTableEntityQuery was itself limited and needs to be
improved a bit. Previously, it always joined the slots table using the
page_latest revision ID, but since it is now used to look up information
for non-latest revisions, it should use the rev_id in that case.
Otherwise, it would return an entity row when called with a MediaInfo ID
and a revision ID that belongs to that entity / page, but where the page
had no entity data in that revision yet.

WikiPageEntityMetaDataLookup::selectRevisionInformationById() is only
called (apart from other WikiPageEntityMetaDataAccessor implementations
delegating their selectRevisionInformationById() calls, and tests) in
one place: WikiPageEntityRevisionLookup. In that class, the call to load
the revision information is immediately followed by an attempt to load
the entity. This means that the change in behavior should be okay for
this caller: it merely means that the attempt to load the entity
(destined to fail anyways) will now be skipped. However, this means we
can now turn the debug log message back into a warning, since we no
longer expect the actual revision loading to fail. The message had been
changed from warn to debug in change I21413771b9 (commit 820d730ea0)
because the warning had started happening frequently in production; I
believe this change represents the better way of avoiding the warning.

Bug: T229482
Bug: T267363
Change-Id: I243bc41b33c790067554b95f37f461072b271710