Tolerate some missing revisions in EntityLoadingHelper
In EntityLoadingHelper::loadEntityRevision, when the
EntityRevisionLookup throws a BadRevisionException, this means that no
entity with that ID is stored in that revision. Instead of immediately
assuming that this means the revision ID is bad, check first whether
that revision could in theory contain that entity ID (i.e. the revision
exists and its title matches the title for that entity ID); if yes,
return null as if the entity had not been found (with no revision ID
parameter). In EntitySavingHelper, remove a check that raises the same
nosuchrevid error for a null entity revision and nonzero base revision
ID (which was always dead code, since EntityLoadingHelper would already
have died with an error in this case), and instead let this case fall
through: if the EntityLoadingHelper returned null (because the baserevid
could in theory contain that entity ID), allow the rest of the method to
potentially create the entity.

In practice, this should only apply to MediaInfo entities: Items,
Properties and Lexemes are stored in the main slot, and it’s impossible
for matching revisions to exist that don’t also contain an entity, and
Senses and Forms can’t be created with custom IDs. For MediaInfo,
though, this should mean that wbeditentity and other APIs can always be
used with a baserevid, even if the mediainfo slot has not yet been
filled.

Bug: T267363
Change-Id: Iab6d598251e72483e0380beb33ab58b9080b62bb