Use EntityTitleStoreLookup in ResultBuilder
ResultBuilder uses the lookup to add page information to a result (page
ID, namespace, last revision ID, etc.), so it should use the interface
that offers the stronger guarantee that the returned title is the one
that stores the entity, not merely an “arbitrary mapping”. Propagate
this difference to other classes that pass the service through to the
ResultBuilder constructor, too.

In practice, there is no difference: WikibaseRepo’s EntityTitleLookup is
already an EntityTitleStoreLookup. (The only non-Store implementation of
EntityTitleLookup lives in WikibaseClient.)

Bug: T267363
Change-Id: I5dcc6c3e1b0486104126a95339f9e062a6821445
Depends-On: Ifd86bc9e19b65a73c9c699271c69e37beedd6e7c