Extract DispatchChangeModificationNotificationJob class
This parent class of DispatchChangeDeletionNotificationJob is analogous
to client’s ChangeModificationNotificationJob, the parent class of
ChangeDeletionNotificationJob and ChangeVisibilityNotificationJob: it
collects some common logic that can be used in two job subclasses – in
this case, finding the entity ID and pushing the jobs to the clients.
(The second job subclass does not yet exist, but will be introduced in
the next commit.)

As far as possible, this commit just moves code around without
introducing too many changes – possible cleanups can be done later.
(`git diff --color-moved` can be useful to verify this.)

Bug: T272334
Change-Id: Ifd59c35c5dde6044241f2d9035282c92e93cac2f
Co-Authored-By: Itamar Givon <itamar.givon@wikimedia.de>