Add DispatchChangeVisibilityNotificationJob
This job is now responsible for pushing ChangeVisibilityNotification
jobs to all client wikis, and is scheduled by the hook handler that used
to (try to) push these jobs directly. This is necessary because pushing
jobs to all client wikis may take too long for a web request.

Like in the parent change, this commit mainly moves code around without
touching it yet. ArticleRevisionVisibilitySetHookHandlerTest is
effectively rewritten, with the old tests all moving over into
DispatchChangeVisibilityNotificationJobTest.

Bug: T272334
Change-Id: Ie8b939a5e86c47212eb247ed25972faa052ad0d9