Stop loading full revisions in DispatchChangeVisibilityNotificationJob
Since the previous change, we no longer need the page of the revision
we’re inspecting; the only piece of information we still need is its
timestamp. RevisionLookup has a dedicated method to load only that
field, which should be rather cheaper than loading the full revision (no
need to join against user, page, comment, etc.). Inline a few methods
and then load only the revision timestamps. (This constructs the
RepoRevisionIdentifiers somewhat earlier than before, since they’re a
convenient container for a revision ID plus timestamp, and we already
know the entity ID as well. An alternative would have been to return an
array of revision ID and timestamp from getEligibleRevisionsById.)

Bug: T272334
Change-Id: Id6175c89b3f6a274414123c356b637a40ca9ffb8