Remove reduntant PHPUnit ->will( … )
Most calls to ->will() can be abbreviated with a more readable form:

* ->will( $this->returnValue( … ) ) becomes ->willReturn( … )
* ->will( $this->returnCallback( … ) ) becomes ->willReturnCallback( … )
* ->will( $this->throwException( … ) ) becomes ->willThrowException( … )
* ->will( $this->returnArgument( … ) ) becomes ->willReturnArgument( … )

There is one subtle difference in the first bullet point, which becomes
relevant in ChangeDispatcherTest. The former form unconditionally
registers the value as the return value of the mocked function, without
checking if it matches the expected return type – if it doesn’t match,
PHP will throw a TypeError when the method is called. With the latter
form, PHPUnit checks that the value is a legal return at the time that
the mock is installed; in the case of ChangeDispatcherTest, where
$siteLinkDiff defaults to null, this means that it will complain that
null is not a valid return value for getSiteLinkDiff(), whereas
previously it would have silently installed that return value and the
method was just never called. Avoid the PHPUnit warning by registering
the return value only if the $siteLinkDiff exists, and otherwise
asserting that the method isn’t called (since we don’t have a valid
return value for it). Likewise, fix the hard-coded return values for
getLatitude() and getLongitude() in GlobeCoordinateDetailsFormatterTest.

Change-Id: Iac24de03f2cf7e64c469472886ea82066ff35845