Use `rawurlencode` to encode external IDs in URLs
This changes url encoding for external IDs to use `rawurlencode` instead
of mediawiki's `wfUrlencode`. This changes the behavior in the following
ways:
* Spaces are now encoded as `%20` instead of `+`.
* The characters `;:@$!*(),~` get encoded. These previously got
  unencoded by `wfUrlencode` to produce prettier URLs for internal title
  links.

As a consequence of the second change in behavior, the test for URLs
formatted as themselves ("$1") is removed. This is not intended to work
and is not currently used on Wikidata according to this SPARQL query:
https://w.wiki/zzZ

Bug: T271126
Change-Id: I2098fb738fc2951d59f46b8d2bc83c4e5cc03441